refactor: Clean up code
